requirements.txt
requests==2.26.0
streamlit==1.20.0
Pillow==9.4.0